% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{annotate_genomic_mutations_with_panel_info}
\alias{annotate_genomic_mutations_with_panel_info}
\title{annotate_genomic_mutations_with_panel_info}
\usage{
annotate_genomic_mutations_with_panel_info(
  genomic_mutations,
  mutation_chromosome_col = "Chromosome",
  mutation_start_col = "Start_position",
  mutation_end_col = "End_position",
  mutation_gene_col = "gene",
  panel_info,
  panel_id_col = "document_type",
  panel_chromosome_col = "chromosome",
  panel_start_col = "start_position",
  panel_end_col = "end_position",
  panel_gene_col = "gene",
  used_panels = c("HP MD Solid Tumor Genomic Assay V1 Report",
    "HP MD Solid Tumor Genomic Assay-DNA 2018 Report",
    "HP MD Liquid Biopsy Panel V1 Report")
)
}
\arguments{
\item{genomic_mutations}{data.frame. wes mutations data.frame.}

\item{mutation_chromosome_col}{chararcter. genomic mutation column for chromosome.}

\item{mutation_start_col}{chararcter. genomic mutation column for start.}

\item{mutation_end_col}{chararcter. genomic mutation column for end.}

\item{mutation_gene_col}{chararcter. genomic mutation column for genename.}

\item{panel_info}{data.frame. information  of panel used for  MDL mutation.}

\item{panel_id_col}{character. column for panel id}

\item{panel_chromosome_col}{chararcter. panel information column for chromosome.}

\item{panel_start_col}{chararcter. panel information column for start.}

\item{panel_end_col}{chararcter. panel information column for end.}

\item{panel_gene_col}{chararcter. panel information column for genename.}

\item{used_panels}{character. panel ids used for current cohort}
}
\value{
data.frame.
}
\description{
annotate genomic mutations (WES/WGS) with provided panel information.
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
