% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{annotate_mdl_mutations_with_mutationdb}
\alias{annotate_mdl_mutations_with_mutationdb}
\title{annotate_mdl_mutations_with_mutationdb}
\usage{
annotate_mdl_mutations_with_mutationdb(
  mdl_mutations,
  mdl_mutations_gene_col = "gene_name",
  mdl_mutations_aa_col = "protein_description",
  with_oncokb = T,
  oncokb_db,
  oncokb_gene_col = "Hugo Symbol",
  with_cosmic = T,
  cosmic_db,
  cosmic_gene_col = "GENE_NAME",
  cosmic_aa_col = "Mutation AA"
)
}
\arguments{
\item{mdl_mutations}{data.frame. mdl mutations data.frame.}

\item{mdl_mutations_gene_col}{chararcter. mdl mutation column for genename.}

\item{mdl_mutations_aa_col}{chararcter. mdl mutation column for amino acid.}

\item{with_oncokb}{logic. annotate with oncoKB or not.}

\item{oncokb_db}{data.frame. oncokb database}

\item{oncokb_gene_col}{character. oncoKB column for genename}

\item{with_cosmic}{logic. annotate with cosmic or not.}

\item{cosmic_db}{data.frame. cosmic database.}

\item{cosmic_gene_col}{character. cosmic column for genename}

\item{cosmic_aa_col}{character. cosmic column for amino acid}
}
\value{
data.frame. mdl mutation with known mutation database.
}
\description{
annotat mdl mutation with known mutation database.
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
