% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{annotate_mdl_mutations_with_timepoints}
\alias{annotate_mdl_mutations_with_timepoints}
\title{annotate_mdl_mutations_with_timepoints}
\usage{
annotate_mdl_mutations_with_timepoints(
  mdl_mutations,
  mdl_mutation_patient_id_col = "Patient_ID",
  mdl_mutation_collect_time_col,
  mdl_mutation_collect_date_format = "\%Y-\%m-\%d",
  timepoint_info,
  timepoint_info_patient_id_col = "Patient_ID",
  timepoint_info_timepoint_col,
  timepoint_info_collect_time_col,
  timepoint_info_collect_date_format = "\%m/\%d/\%Y"
)
}
\arguments{
\item{mdl_mutations}{data.frame. mdl mutation data.frame.}

\item{mdl_mutation_patient_id_col}{character. mdl mutation column for patient id}

\item{mdl_mutation_collect_time_col}{character. mdl mutation column for collect time}

\item{mdl_mutation_collect_date_format}{character. mdl mutation collect time column date format.}

\item{timepoint_info}{data.frame. mdl sample timepoint information}

\item{timepoint_info_patient_id_col}{character. mdl sample timepoint information column for patient id}

\item{timepoint_info_timepoint_col}{character. mdl sample timepoint information column for timepoint}

\item{timepoint_info_collect_time_col}{character. mdl sample timepoint information column for collect time.}

\item{timepoint_info_collect_date_format}{character. time format for timepoint info collect time.}
}
\value{
data.frame. timepoint annotated mdl mutations
}
\description{
annotate mdl mutations with timepoints data. mark sample as well as mutations with time points(like "Baseline","TP2",...) according to the minimal gap between collect time.
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
