% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{cancerhallmark_heatmap}
\alias{cancerhallmark_heatmap}
\title{cancerhallmark_heatmap
heatmap of cancer hallmarks}
\usage{
cancerhallmark_heatmap(
  assay,
  assay_format = "l",
  gene_id_column = "Hugo_Symbol",
  sample_id_col = "Sample_ID",
  value_col,
  cancerhallmarks = CosmicCancerGeneCensusHallmarks,
  top_anno,
  top_anno_sample_order,
  hallmark_colors = NULL,
  impact_colors = NULL,
  show_assay_color = F,
  assay_colors = NULL,
  ...
)
}
\arguments{
\item{assay}{data.frame. assay data, such as mutations.}

\item{assay_format}{character. format of assay, like "l" for long format}

\item{gene_id_column}{character. column for gene id.}

\item{sample_id_col}{character. column for sample id.}

\item{value_col}{character. column for value.}

\item{cancerhallmarks}{data.frame. cancer hallmarks.}

\item{top_anno}{HeatmapAnnotation. top annotation.}

\item{top_anno_sample_order}{character. sample order of top annotation.}

\item{hallmark_colors}{character. colors for hallmarks.}

\item{impact_colors}{character. colors for impact.}

\item{show_assay_color}{logic. whether show assay color individually.}

\item{assay_colors}{character. colors for assay value.}

\item{...}{list. parameter passed to Heatmap.}
}
\value{
list. contains hallmark_heatmap and underlying matrix.
}
\description{
cancerhallmark_heatmap
heatmap of cancer hallmarks
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
