% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_immune_deconvolution.R
\name{CIBERSORT}
\alias{CIBERSORT}
\title{CIBERSORT
Main functions}
\usage{
CIBERSORT(X, Y, perm = 0, QN = TRUE)
}
\arguments{
\item{X}{matrix. ig_matrix file path to gene expression from isolated cells}

\item{Y}{matrix. mixture_file heterogenous mixed expression}

\item{perm}{numeric. perm Number of permutations}

\item{QN}{logic. Perform quantile normalization or not (TRUE/FALSE)}
}
\description{
CIBERSORT
Main functions
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mRNA.R (2026-03-02).
}
