% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{complex_oncoplot_o}
\alias{complex_oncoplot_o}
\title{complex_oncoplot_o
oncoplot of mutation and cnv}
\usage{
complex_oncoplot_o(
  snp_indels,
  for_row = "Hugo_Symbol",
  for_column = "Tumor_Sample_Barcode",
  for_value = "Variant_Classification",
  cnv = TRUE,
  cnv_types = c("Amplification", "Deletion"),
  annotate_source = F,
  source_column = "Source",
  source_label = c(MDL = "*"),
  sample_order = NULL,
  selected_genes,
  gene_order,
  mutsigCV2_sig_genes = NA,
  mutsigCV2_sig_p_threshold = 0.05,
  multi_hit = FALSE,
  multi_hit_col = "black",
  top_filter = 20L,
  col,
  bg_color = "#dcdedc",
  remove_empty_sample = F,
  remove_macromolecular_gene = F,
  mark_macromolecular_gene = F,
  macromolecular_mark = "*",
  macromolecular_threshold = 10000,
  top_annotation = NULL,
  top_anno_sample_order = NULL,
  output_dir = "./",
  ...
)
}
\arguments{
\item{snp_indels}{data.frame. longer_style snp and indels (only nonSilent), and also cnv data, should be provided without default, column of sample id should be factor.}

\item{for_row}{character. col used for row names of oncoplot data,default value: Hugo_Symbol.}

\item{for_column}{character. col used for column names of oncoplot data,default value: Tumor_Sample_Barcode.}

\item{for_value}{character. col used for value of oncoplot data,default value: Variant_Classification.}

\item{cnv}{logic. include cnv into the oncoplot,default TRUE.}

\item{cnv_types}{character vector. if cnv is TRUE, what kinds of alteration would be regarded as cnv}

\item{annotate_source}{logic. hether to mark mutation source.}

\item{source_column}{character. column for source.}

\item{source_label}{character. label for source.}

\item{sample_order}{character. sample order.}

\item{selected_genes}{character. selected genes that will be shown.}

\item{gene_order}{character. gene order.}

\item{mutsigCV2_sig_genes}{data.frame. significant genes from mutsig2CV.}

\item{mutsigCV2_sig_p_threshold}{numeric. threshold used to filter significant driver genes.}

\item{multi_hit}{logic. combine various mutation and cnv as multi_hit, default FALSE.}

\item{multi_hit_col}{character. color for multi_hit, default black.}

\item{top_filter}{numeric. number of top frequent genes would be plotted,default 20.}

\item{col}{character.color pallete for oncoplot, default NA that oncoplot will use its pallete.}

\item{bg_color}{character.background color, default grey (#' dcdedc).}

\item{remove_empty_sample}{logic. remove empty samples, default FALSE.}

\item{remove_macromolecular_gene}{logic. remove macromolecular genes that detemined by macromolecular_threshold, default FALSE.}

\item{mark_macromolecular_gene}{logic. mark macromolecular genes with macromolecular_mark.}

\item{macromolecular_mark}{character. used to mark macromolecular gene if set remove_macromolecular_gene as FALSE, default *.}

\item{macromolecular_threshold}{numeric. macromolecular threshold that determine whether macromolecular, default 10000bp.}

\item{top_annotation}{HeatmapAnotation. top annotation for oncoplot, default NULL, if NULL, total alteration burden gona be used as top annotaion.}

\item{top_anno_sample_order}{character. sample orders for top annotation.}

\item{output_dir}{character. output plot directory.}

\item{...}{list. other params used for oncoprint.}
}
\value{
data.frame. mutations.
}
\description{
complex_oncoplot_o
oncoplot of mutation and cnv
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
