% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_clustering.R
\name{consensusCluster}
\alias{consensusCluster}
\title{consensusCluster}
\usage{
consensusCluster(
  data,
  scale = c("row", "column", "none"),
  method = c("bootstrap", "perturb", "combine"),
  subFeatureSize = 0.8,
  subSampleSize = 1,
  noise = 1,
  cutFUN,
  nTimes = 100,
  clusters = 2,
  verbose = F,
  ...
)
}
\arguments{
\item{data}{\code{data.frame()} or \code{matrix()}. Gene expression data.}

\item{scale}{\code{character()}. Default ="row", Direction for scale data.}

\item{method}{\code{character()}. default="bootstrap",  method used to produce subset expression for clustering}

\item{subFeatureSize}{\code{numeric()}. Default=0.8. The subset fraction of features for bootstrap.}

\item{subSampleSize}{\code{numeric()}. Default=1. The subset fraction of samples for bootstrap.}

\item{noise}{\code{numeric()}. Default=1, noise with variance of noise defined unit will be added onto expression}

\item{cutFUN}{\code{function()}. A vector of clustering functions. Options include:
\itemize{
\item "Methods in ClassDiscovery package" - "cutHclust", "cutKmeans", "cutPam", and "cutRepeatedKmeans"
\item "Methods provided by expr" - "kmeansCluster", "pamCluster", "hclustCluster", "fuzzyCluster", "mclustCluster", "apclustCluster", "hdbscanCluster", "mclCluster", "speccCluster", "kkmeansCluster", "skmeansCluster", "nmfCluster", "somCluster"
}}

\item{nTimes}{\code{integer()}. Default=100. Times of iteration of Clustering}

\item{clusters}{\code{vector(mode = "integer")}. Default=2, predefined number of clusters.}

\item{verbose}{\code{logical()}. default \code{FALSE}, if \code{TRUE}, print detailed process information.}

\item{...}{params passed on to cutFUN.}
}
\value{
\code{data.frame()}, containing possibility of co-clustering.
}
\description{
consensus clustering using bootstrap or(and) noise addition of expression.
}
\details{
Source provenance: GBM/XL_figures/vDotJun0823/GBMnat/R/funcsGBMnat.R (2025-08-11).
}
\examples{
\dontrun{
results<-future_consensusCluster(data,method='combine',clusters=4)
}
}
