% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{convert_mutations}
\alias{convert_mutations}
\title{convert_mutations}
\usage{
convert_mutations(
  mutations,
  target = c("mutsig2cv", "maftools", "pynbs"),
  msc_snp_conversion_dict,
  msc_indel_conversion_dict,
  maftools_mutation_conversion_dict,
  pynbs_nonSilent_variant_classication
)
}
\arguments{
\item{mutations}{data.frame. mutations data.frame including SNPs and Indels.}

\item{target}{character. targeted algorithm.}

\item{msc_snp_conversion_dict}{character. a named character mapping vector between cgl snp and mutsig2cv snp.}

\item{msc_indel_conversion_dict}{character. a named character mapping vector between cgl indel and mutsig2cv indel.}

\item{maftools_mutation_conversion_dict}{character. a named character mapping vector between cgl snp/indel and maftools snp/indel.}

\item{pynbs_nonSilent_variant_classication}{character. nonSilent variant classification for pyNBS.}
}
\value{
data.frame. converted data.frame fit specifid algorithm
}
\description{
convert mutations used for mutsig2cv, maftools and pynbs
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
