% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_immune_deconvolution.R
\name{CoreAlg}
\alias{CoreAlg}
\title{CoreAlg
Core algorithm}
\usage{
CoreAlg(X, y)
}
\arguments{
\item{X}{matrix. cell-specific gene expression}

\item{y}{matrix. mixed expression per sample}
}
\description{
CoreAlg
Core algorithm
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mRNA.R (2026-03-02).
}
