% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_immune_deconvolution.R
\name{DeconRNASeq_}
\alias{DeconRNASeq_}
\title{DeconRNASeq_}
\usage{
DeconRNASeq_(
  datasets,
  signatures,
  proportions = NULL,
  checksig = FALSE,
  known.prop = FALSE,
  use.scale = TRUE,
  fig = TRUE
)
}
\arguments{
\item{datasets}{data.frame. expression data.frame}

\item{signatures}{data.frame. reference signature data.frame}

\item{proportions}{data.frame. proportion matrix from different tissue/cell types.}

\item{checksig}{logic. whether the condition number of signature matrix should be checked, default = FALSE}

\item{known.prop}{logic. whether the proportions of cell types have been known in advanced for proof of concept, default = FALSE}

\item{use.scale}{logic. whether the data should be centered or scaled, default = TRUE}

\item{fig}{logic. whether to generate the scatter plots of the estimated cell fractions vs. the true proportions of cell types, default = TRUE}
}
\value{
list. a list containing estimated cell type fraction matrix for all the mixture samples and svd calculated PCA on the mixture samples to estimate the number of pure sources according to the cumulative R2
}
\description{
a replacement for DeconRNASeq, fix some bugs
}
