% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_expression_analysis.R
\name{dge_DESeq}
\alias{dge_DESeq}
\title{dge_DESeq
differential gene expression analysis for raw count expressions using DESeq2 algorithm}
\usage{
dge_DESeq(
  expressions,
  sample_frequency_threshold = 0.5,
  clinic_info,
  ID_col,
  group_col,
  covariate_col,
  block_col,
  contrasts
)
}
\arguments{
\item{expressions, }{dataframe, no default, gene expressions table,must be provided}

\item{sample_frequency_threshold, }{numeric, default 0.5, genes with low expression occurred in at least sample_frequency_threshold fraction of all samples will be removed}

\item{clinic_info, }{dataframe, no default, clinic information table, must be provide}

\item{ID_col, }{character, no default, column name for Sample ID (or Patient ID), should be consistent with column names of expressions}

\item{group_col, }{character, no default, column name for group factor that differential gene expression will be conducted based on}

\item{covariate_col, }{character, no default, column name for covariate factor that effects should be removed from model}

\item{block_col, }{character, no default, column name for block factor if test is conducted on block model, such as paired test}

\item{contrasts, }{string vector, no default, specific contrasts if prefered, elements should be exactly same as group factor}
}
\value{
list. contains expressions,methond, design,contrasts,test, and statistics of DESeq2 test
}
\description{
dge_DESeq
differential gene expression analysis for raw count expressions using DESeq2 algorithm
}
\details{
Source provenance: mRNA.R (2026-03-02).
}
