% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_expression_analysis.R
\name{dge_limma}
\alias{dge_limma}
\title{Differential expression analysis using limma.}
\usage{
dge_limma(
  expressions,
  is_rawcount = FALSE,
  is_logged = T,
  normalize = FALSE,
  sample_frequency_threshold = 0.5,
  clinic_info,
  ID_col,
  group_col,
  covariate_col,
  block_col,
  contrasts,
  method = c("limma_trend", "limma_voom"),
  robust_lm = FALSE,
  weight = FALSE
)
}
\arguments{
\item{expressions}{\code{data.frame()}. Gene expressions table.}

\item{is_rawcount}{\code{logical()}. Whether expressions is raw count or not. Default set to FALSE.}

\item{is_logged}{\code{logical()}. Whether provided expressions was log2 transformed or not. Default set to TRUE.}

\item{normalize}{\code{logical()}. Normalize provided expressions by library size. Default set to FALSE.}

\item{sample_frequency_threshold}{\code{numeric()}. Genes with low expression occurred in at least sample_frequency_threshold fraction of all samples will be removed. Default 0.5.}

\item{clinic_info}{\code{data.frame()}. Clinical information table.}

\item{ID_col}{\code{character()}. Column name for Sample ID (or Patient ID), should be consistent with column names of expressions.}

\item{group_col}{\code{character()}. Column name for group factor that differential gene expression will be conducted on.}

\item{covariate_col}{\code{character()}. column name for covariate factor that effects should be removed from model.}

\item{block_col}{\code{character()}. Column name for block factor if test is conducted on block model, such as paired test.}

\item{contrasts}{\code{vector(mode="character")} Specific contrasts if preferred, elements should be exactly same as group factor.}

\item{method}{\code{character(1)}. Method to conduct test. One of:
\itemize{
\item "limma_trend" for non raw count expressions
\item "limma_voom" for raw count expressions.
}
Default set as "limma_trend".}

\item{robust_lm}{\code{logical(1)}. Whether to use robust linear model. Default set to FALSE.}

\item{weight}{\code{logical(1)}. Whether to use weights in linear model. Default set to FALSE.}
}
\value{
\code{list()}, contains expressions, method, design, contrasts, test, and statistics of limma test.
}
\description{
differential gene expression analysis using limma algorithm.
}
\details{
Source provenance: funcsInPembro.R (2025-07-02).
}
\examples{
\dontrun{
data(woodman)
clean_log2_protein_expressions=log2_expressions
results<-dge_limma(
  clean_log2_protein_expressions,
  clinical_info = clean_RNA_sample_info,
  ID_col = "Sample_ID",
  group_col = "APOLLO_TIMPOINT",
  contrasts = c("TP2-Baseline"),
  method ="limma_trend")
}


}
