% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_immune_deconvolution.R
\name{doPerm}
\alias{doPerm}
\title{doPerm
do permutations}
\usage{
doPerm(perm, X, Y)
}
\arguments{
\item{perm}{numeric. Number of permutations.}

\item{X}{matrix. cell-specific gene expression.}

\item{Y}{matrix. mixed expression per sample.}
}
\description{
doPerm
do permutations
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mRNA.R (2026-03-02).
}
