% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_clustering.R
\name{estimate_bestNumberofClusters}
\alias{estimate_bestNumberofClusters}
\title{estimate best number of clusters}
\usage{
estimate_bestNumberofClusters(
  data,
  scale = c("row", "column", "none"),
  data_preTreat = T,
  removeVar = 0.2,
  variance_explained_cutoff = 0.8,
  method = "complete",
  min.nc = 2,
  max.nc = 15,
  plotOP = FALSE
)
}
\arguments{
\item{data}{\code{data.frame()} or \code{matrix()}. Gene expression data (or matrix).}

\item{scale}{\code{character()}. Direction for scale data. Default ="row".}

\item{data_preTreat}{\code{logical()}. If TRUE, data will be PCA transformed; Components whose cumulative sum of variance reach variance_explained_cutoff will be kept. Default FALSE.}

\item{removeVar}{\code{numeric()}. Remove this percentage of variables based on low variance. Default=0.2.}

\item{variance_explained_cutoff}{\code{numeric()}. If data_preTreat is TRUE, components in PCA transformed data matrix whose cumulative sum of variance should reach this value. default=0.8.}

\item{method}{\code{character()}.The clustering method to be used. This should be one of: "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid", "kmeans". Default = "complete".}

\item{min.nc}{\code{numeric()}. Minimal number of clusters, between 1 and (number of objects - 1), default=2.}

\item{max.nc}{\code{numeric()}. Maximal number of clusters, between 2 and (number of objects - 1), greater or equal to min.nc. By default, max.nc=15.}

\item{plotOP}{\code{logical()}. Whether to output plots.}
}
\value{
\code{list()} containing \code{Best.nc()} (statistics for various indices) and \code{Best.NumberofCluster()} (best number of clusters).
}
\description{
estimate the best number of cluster for clustering, using 30 indices from NbClust::NbClust
}
\details{
Source provenance: GBM/XL_figures/vDotJun0823/GBMnat/R/funcsGBMnat.R (2025-08-11).
}
\examples{
\dontrun{
result<-estimate_bestNumberofClusters(data=data,scale='none',data_preTreat = F,max.nc = 8)
}
}
