% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_statistics_utilities.R
\name{filter_assay_info}
\alias{filter_assay_info}
\title{Filter assay info by sample list}
\usage{
filter_assay_info(assay_info, sample_info)
}
\arguments{
\item{assay_info}{Input data object used by the function.}

\item{sample_info}{Input data object used by the function.}
}
\value{
A transformed data object returned by the function.
}
\description{
Filter assay info by sample list
}
\details{
Source provenance: functions.R (2026-03-02).
}
\examples{
\dontrun{
filter_assay_info(assay_info = ..., sample_info = ...)
}
}
