% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{filter_mutects}
\alias{filter_mutects}
\title{filter_mutects}
\usage{
filter_mutects(
  mutects,
  traced_columns = c("Sample_ID", "gene", "type", "chr", "start", "end"),
  low_purity_filter = T,
  low_purity_samples,
  treat_noisy_samples = F,
  noisy_samples = c(),
  exonutronly_filter = T,
  keep = c("exonic", "exonic;splicing", "splicing", "UTR3", "UTR5", "UTR5;UTR3"),
  off_target_filter = T,
  intervals = intervals,
  germline_filter = T,
  cutoff_n_vaf = 0.01,
  treat_known_cancer_gene_specially = T,
  known_cancer_genes = c("TP53", "GNA11", "GNAQ"),
  special_mutect_genes = c(),
  mapping_quality_filter = T,
  cutoff_mq = 30,
  statistical_filter = T,
  cutoff_prop_test_p = 0.01,
  tumor_f_filter = T,
  cutoff_tumor_f = 0,
  tumor_coverage_filter = T,
  cutoff_t_coverage = 10,
  normal_coverage_filter = T,
  cutoff_n_coverage = 10,
  lodt_filter = T,
  cutoff_lodt = 6.3,
  consistent_mutect_statistic_filter = T,
  mutect_for_consistent_statistic = NA,
  cutoff_binom_pval = 0.05,
  description_file = "description.txt",
  save_traced_mutect = TRUE,
  traced_mutect_file = "traced_mutect.txt"
)
}
\arguments{
\item{mutects}{dataframe. containing snv annotation.}

\item{traced_columns}{character. default c("Sample_ID","gene","type","chr","start","end"),used for removing dupicated snv annotations.}

\item{low_purity_filter}{logic. default TRUE, remove low purity sample or keep.}

\item{low_purity_samples}{character. list of samples with low purity.}

\item{treat_noisy_samples}{logic. default TRUE, remove specially treat noisy samples (remove all mutects not present in clean/rest samples).}

\item{noisy_samples}{character. list of samples with too much noise.}

\item{exonutronly_filter}{logic. default TRUE, keep only snv located in exon, utr or not.}

\item{keep}{character. default c("exonic","exonic;splicing","splicing","UTR3","UTR5","UTR5;UTR3"), list snv locus (column func.knowngene) to keep.}

\item{off_target_filter}{logic. default TRUE, remove off-target snv or not.}

\item{intervals}{GenomicRanges.providing the targeted regions information.}

\item{germline_filter}{logic. default TRUE, remove germline snv or not.}

\item{cutoff_n_vaf}{numeric. defalut 0.01, cutoff for normal variant allele fraction.}

\item{treat_known_cancer_gene_specially}{logic. defalut TRUE, split known cancer genes and treat differently.}

\item{known_cancer_genes}{character. list of cancer gene names.}

\item{special_mutect_genes}{character. list of genes whose snv not through filter sets.}

\item{mapping_quality_filter}{logic. default TRUE, remove low mapping quality snv or not.}

\item{cutoff_mq}{numeric. default 30, cuttoff for maximal mapping quality of reads.}

\item{statistical_filter}{logic.default TRUE, remove snv not passing proportional test or not.}

\item{cutoff_prop_test_p}{numeric. default 0.01, cutoff for proportial test pvalue.}

\item{tumor_f_filter}{logic. default TRUE, remove snv with low tumor fraction or not.}

\item{cutoff_tumor_f}{numeric. default 0.0, cufoff for tumor fraction.}

\item{tumor_coverage_filter}{logic. default TRUE, remove snv with low tumor coverage or not.}

\item{cutoff_t_coverage}{numeric. default 10, cutoff for tumor coverage (column t_alt+ column t_ref).}

\item{normal_coverage_filter}{logic. default TRUE, remove snv with low normal coverage or not.}

\item{cutoff_n_coverage}{numeric. default 10, cutoff for normal coverage (column n_alt+ column n_ref).}

\item{lodt_filter}{logic. default TRUE, remove snv with low log likelyhood of tumor event or not.}

\item{cutoff_lodt}{numeric. default 6.3, cutoff for number of log likelyhood of tumor event.}

\item{consistent_mutect_statistic_filter}{logic. default TRUE, remove snv failed to pass consistency statistical test (binom test).}

\item{mutect_for_consistent_statistic}{data.frame. containing gene,number of snv, number of samples (columns: gene,n_pindel,n_sample), used for binom test with targeted snv.}

\item{cutoff_binom_pval}{numeric. default 0.05, cutoff for binom test pvalue.}

\item{description_file}{character. default description.txt, file name for deposit of information of each operation (filter sets).}

\item{save_traced_mutect}{logic. default TRUE, save traced_mutect_file.}

\item{traced_mutect_file}{character. default traced_pindel.txt, file name for deposit results after each operation (filter sets) if save_traced_mutect=TRUE.}
}
\value{
data.frame. filtered mutects(snp).
}
\description{
filter mutects using various filter sets
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
