% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{filter_pindels}
\alias{filter_pindels}
\title{filter_pindels}
\usage{
filter_pindels(
  pindels,
  traced_columns = c("Sample_ID", "gene", "type", "chr", "start", "end"),
  low_purity_filter = T,
  low_purity_samples,
  treat_noisy_samples = F,
  noisy_samples = c(),
  exonutronly_filter = T,
  keep = c("exonic", "exonic;splicing", "splicing", "UTR3", "UTR5", "UTR5;UTR3"),
  remove_td_filter = T,
  td_coverage_filter = T,
  cutoff_td_coverage = 1.5,
  off_target_filter = T,
  intervals = intervals,
  germline_filter = T,
  cutoff_n_vaf = 0.01,
  treat_known_cancer_gene_specially = T,
  known_cancer_genes = c("TP53", "GNA11", "GNAQ"),
  special_pindel_genes = c("BAP1", "PRKDC"),
  mapping_quality_filter = T,
  cutoff_mq = 25,
  statistical_filter = T,
  cutoff_prop_test_p = 0.01,
  t_vaf_filter = T,
  cutoff_t_vaf = 0,
  tumor_f_filter = T,
  cutoff_tumor_f = 0,
  deletion_length_filter = T,
  cutoff_d_length = 1000,
  insertion_length_filter = T,
  cutoff_i_length = 6,
  tumor_coverage_filter = T,
  cutoff_t_coverage = 10,
  normal_coverage_filter = T,
  cutoff_n_coverage = 10,
  support_filter = T,
  cutoff_support = 3,
  homopolymer_filter = T,
  maxn_homopolymer = 6,
  repeat_pindel_filter = T,
  cutoff_repeat = 2,
  consistent_pindel_statistic_filter = T,
  pindel_for_consistent_statistic = NA,
  cutoff_binom_pval = 0.05,
  description_file = "description.txt",
  save_traced_pindel = T,
  traced_pindel_file = "traced_pindel.txt"
)
}
\arguments{
\item{pindels}{dataframe. containing insertion and deletion annotation.}

\item{traced_columns}{character. default c("Sample_ID","gene","type","chr","start","end"),used for removing dupicated pindel annotations.}

\item{low_purity_filter}{logic. default TRUE, remove low purity sample or keep.}

\item{low_purity_samples}{character. list of samples with low purity.}

\item{treat_noisy_samples}{logic. default TRUE, remove specially treat noisy samples (remove all pindels not present in clean/rest samples).}

\item{noisy_samples}{character. list of samples with too much noise.}

\item{exonutronly_filter}{logic. default TRUE, keep only pindels located in exon, utr or not.}

\item{keep}{character. default c("exonic","exonic;splicing","splicing","UTR3","UTR5","UTR5;UTR3"), list pindel locus (column func.knowngene) to keep.}

\item{remove_td_filter}{logic. default TRUE, remove tandem duplicated or not.}

\item{td_coverage_filter}{logic. default TRUE, remove low coverage tandom duplicate.}

\item{cutoff_td_coverage}{numeric. default 1.5, cutoff for ratio of tandem duplicate to average genome coverage.}

\item{off_target_filter}{logic. default TRUE, remove off-target pindels or not.}

\item{intervals}{GenomicRanges.providing the targeted regions information.}

\item{germline_filter}{logic. default TRUE, remove germline pindels or not.}

\item{cutoff_n_vaf}{numeric. defalut 0.01, cutoff for normal variant allele fraction.}

\item{treat_known_cancer_gene_specially}{logic. defalut TRUE, split known cancer genes and treat differently.}

\item{known_cancer_genes}{character. list of cancer gene names.}

\item{special_pindel_genes}{character. list of genes whose pindel no through filter sets.}

\item{mapping_quality_filter}{logic. default TRUE, remove low mapping quality pindels or not.}

\item{cutoff_mq}{numeric. default 25, cuttoff for averaged mapping quality of anchor reads.}

\item{statistical_filter}{logic.default TRUE, remove pindels not passing proportional test or not.}

\item{cutoff_prop_test_p}{numeric. default 0.01, cutoff for proportial test pvalue.}

\item{t_vaf_filter}{logic. default TRUE, remove pindels with low tumor variant allele fraction or not.}

\item{cutoff_t_vaf}{numeric. default 0.0, cutoff for tumor variant allele fraction.}

\item{tumor_f_filter}{logic. default TRUE, remove pindels with low tumor fraction or not.}

\item{cutoff_tumor_f}{numeric. default 0.0, cufoff for tumor fraction.}

\item{deletion_length_filter}{logic. defalut TRUE, remove too long deletions or not.}

\item{cutoff_d_length}{numeric. default 1000, cutoff for deletion length.}

\item{insertion_length_filter}{logic. defalut TRUE, remove too short insertions or not.}

\item{cutoff_i_length}{numeric. default 6, cutoff for insertiontion length.}

\item{tumor_coverage_filter}{logic. default TRUE, remove pindels with low tumor coverage or not.}

\item{cutoff_t_coverage}{numeric. default 10, cutoff for tumor coverage (column t_alt+ column t_ref).}

\item{normal_coverage_filter}{logic. default TRUE, remove pindels with low normal coverage or not.}

\item{cutoff_n_coverage}{numeric. default 10, cutoff for normal coverage (column n_alt+ column n_ref).}

\item{support_filter}{logic default TRUE, remove pindels with low support reads or not.}

\item{cutoff_support}{numeric. default 3, cutoff for number of support reads.}

\item{homopolymer_filter}{logic. default TRUE, remove pindels with too many homopolymer or not.}

\item{maxn_homopolymer}{numeric. default 6, allowed maximal numer of homopolymer.}

\item{repeat_pindel_filter}{logic. default TRUE, remove repeated pindels or not.}

\item{cutoff_repeat}{numeric. default 2, cutoff for repeating frequency.}

\item{consistent_pindel_statistic_filter}{logic. default TRUE, remove pindels failed to pass consistency statistical test (binom test).}

\item{pindel_for_consistent_statistic}{data.frame. containing gene,number of pindels, number of samples (columns: gene,n_pindel,n_sample), used for binom test with targeted pindels.}

\item{cutoff_binom_pval}{numeric. default 0.05, cutoff for binom test pvalue.}

\item{description_file}{character. default description.txt, file name for deposit of information of each operation (filter sets).}

\item{save_traced_pindel}{logic. default TRUE, save traced_pindel_file.}

\item{traced_pindel_file}{character. default traced_pindel.txt, file name for deposit results after each operation (filter sets) if save_traced_pindel=TRUE.}
}
\value{
data.frame. filtered pindels.
}
\description{
filter pindels using various filter sets
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
