% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{get_combined_tcga_mutation_table}
\alias{get_combined_tcga_mutation_table}
\title{get_combined_tcga_mutation_table
combine tumor specific tcga mutation simplified table with pre-existing non-tumor specific tcga mutation simplified table}
\usage{
get_combined_tcga_mutation_table(
  tcga_tumor_abbreviation = "",
  tcga_studies,
  tcga_snp_table,
  tcga_indel_table
)
}
\arguments{
\item{tcga_tumor_abbreviation}{character. abbreviation for tcga tumor.}

\item{tcga_studies}{data.frame. TCGA study list.}

\item{tcga_snp_table}{data.frame. simplified TCGA snp frequency table.}

\item{tcga_indel_table}{data.frame. simplified TCGA ins/del frequency table.}
}
\value{
list. contain combined simplified tcga snp/indel frequency table
}
\description{
get_combined_tcga_mutation_table
combine tumor specific tcga mutation simplified table with pre-existing non-tumor specific tcga mutation simplified table
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
