% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{get_min_multihit_distance}
\alias{get_min_multihit_distance}
\title{get_min_multihit_distance}
\usage{
get_min_multihit_distance(codons)
}
\arguments{
\item{codons}{numeric. a numeric vector describ the protein codons}
}
\value{
numeric. minimal distance among codons.
}
\description{
calculte the minimal distance for multiple mutations
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
