% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_survival_analysis.R
\name{getKM_medium}
\alias{getKM_medium}
\title{get KM medium survival time}
\usage{
getKM_medium(
  survdf,
  duration,
  variable = NULL,
  censoring,
  conversion = "d2m",
  timeUnit = "month"
)
}
\arguments{
\item{survdf}{data frame with survival data.}

\item{duration}{name of the column with the duration of the event.}

\item{variable}{name of the column with the variable to stratify the analysis.}

\item{censoring}{name of the column with the censoring information.}

\item{conversion}{"d2m" for days to months conversion.}

\item{timeUnit}{time unit for the duration column.}
}
\value{
a data frame with the medium survival time for each group.
}
\description{
get KM medium survival time
}
\details{
Source provenance: GBM/XL_figures/vDotJun0823/GBMnat/R/funcsGBMnat.R (2025-08-11).
}
