% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{harmonize_mutations}
\alias{harmonize_mutations}
\title{harmonize_mutations}
\usage{
harmonize_mutations(
  wes_sample_info,
  wes_sample_id_col,
  wes_patient_id_col,
  wes_timepoint_col,
  wes_sample_collect_col,
  wes_mutations = NULL,
  wes_mutation_sample_id_col,
  wes_gene_col = "gene",
  wes_chromosome_col = "Chromosome",
  wes_start_col = "Start_Position",
  wes_end_col = "End_Position",
  wes_keep_columns = c("Variant_Classification", "Variant_Type", "Reference_Allele",
    "Tumor_Seq_Allele1", "Tumor_Seq_Allele2", "Protein_Change", "i_TumorVAF_WU", "cdna"),
  wes_seq_intervals,
  wgs_sample_info,
  wgs_sample_id_col,
  wgs_patient_id_col,
  wgs_timepoint_col,
  wgs_sample_collect_col,
  wgs_mutations = NULL,
  wgs_mutation_sample_id_col,
  wgs_gene_col = "gene",
  wgs_chromosome_col = "Chromosome",
  wgs_start_col = "Start_Position",
  wgs_end_col = "End_Position",
  wgs_keep_columns = c("transcript", "aaannotation", "func.knowngene",
    "exonicfunc.knowngene", "cdna", "chrom", "start", "end", "ref_allele", "alt_allele",
    "t_ref_count", "t_alt_count", "n_ref_count", "n_alt_count"),
  mdl_sample_info,
  mdl_sample_id_col = "Sample_ID",
  mdl_patient_id_col = "Patient_ID",
  mdl_timepoint_col,
  mdl_sample_collect_time_col,
  mdl_sample_collect_date_format = "\%Y-\%m-\%d",
  mdl_mutations = NULL,
  mdl_mutation_sample_id_col = NULL,
  mdl_mutation_patient_id_col = "Patient_ID",
  mdl_mutation_collect_time_col,
  mdl_mutation_collect_date_format = "\%Y-\%m-\%d",
  mdl_mutation_gene_col = "gene",
  mdl_mutation_indication_col = "indication_name",
  mdl_mutations_aa_col = "protein_description",
  mdl_mutation_panel_col = "document_subtype_description",
  mdl_mutation_keep_columns = c("aa_change", "indication_name", "nucleotide_change",
    "reference_nucleotide", "alternate_nucleotide", "nucleotide_change_type",
    "nucleotide_change_subtype"),
  panel_info,
  panel_id_col = "document_type",
  panel_chromosome_col = "Chromosome",
  panel_start_col = "Start_Position",
  panel_end_col = "End_Position",
  panel_gene_col = "gene",
  with_oncokb = T,
  oncokb_db,
  oncokb_gene_col = "Hugo Symbol",
  with_cosmic = T,
  cosmic_db,
  cosmic_gene_col = "GENE_NAME",
  cosmic_aa_col = "Mutation AA",
  mdl_columns2wes_columns = c(aa_change = "Protein_Change", nucleotide_change = "cdna",
    reference_nucleotide = "Reference_Allele", alternate_nucleotide =
    "Tumor_Seq_Allele2"),
  combined = F
)
}
\arguments{
\item{wes_sample_info}{data.frame. WES sample information.}

\item{wes_sample_id_col}{character. WES sample information column for sample id.}

\item{wes_patient_id_col}{character. WES sample information column for patient id.}

\item{wes_timepoint_col}{character. WES sample information column for timepoint.}

\item{wes_sample_collect_col}{character. WES sample information column for collect time.}

\item{wes_mutations}{data.frame. WES mutations.}

\item{wes_mutation_sample_id_col}{character. WES mutation column for sample id.}

\item{wes_gene_col}{character. WES mutation column for genename.}

\item{wes_chromosome_col}{character. WES mutation column for chromosome.}

\item{wes_start_col}{character. WES mutation column for start.}

\item{wes_end_col}{character. WES mutation column for end.}

\item{wes_keep_columns}{character. WES mutation column kept for the output.}

\item{wes_seq_intervals}{grangeslist. intervals for WES sequencing.}

\item{wgs_sample_info}{data.frame. WGS sample information.}

\item{wgs_sample_id_col}{character. WGS sample information column for sample id.}

\item{wgs_patient_id_col}{character. WGS sample information column for patient id.}

\item{wgs_timepoint_col}{character. WGS sample information column for timepoint.}

\item{wgs_sample_collect_col}{character. WGS sample information column for collect time.}

\item{wgs_mutations}{data.frame. WGS mutations.}

\item{wgs_mutation_sample_id_col}{character. WGS mutation column for sample id.}

\item{wgs_gene_col}{character. WGS mutation column for genename.}

\item{wgs_chromosome_col}{character. WGS mutation column for chromosome.}

\item{wgs_start_col}{character. WGS mutation column for start.}

\item{wgs_end_col}{character. WGS mutation column for end.}

\item{wgs_keep_columns}{character. WGS mutation column kept for the output.}

\item{mdl_sample_info}{data.frame. mdl sample information.}

\item{mdl_sample_id_col}{character. mdl sample information column for sample id.}

\item{mdl_patient_id_col}{character. mdl sample information column for patient id.}

\item{mdl_timepoint_col}{character. mdl sample information column for timepoint.}

\item{mdl_sample_collect_time_col}{character. mdl sample information column for collect time.}

\item{mdl_sample_collect_date_format}{character. timepoint information collect time column date format.}

\item{mdl_mutations}{data.frame. mdl mutations.}

\item{mdl_mutation_sample_id_col}{character. mdl mutation column for sample id.}

\item{mdl_mutation_patient_id_col}{character. mdl mutation column for patient id.}

\item{mdl_mutation_collect_time_col}{character. mdl sample information column for collect time.}

\item{mdl_mutation_collect_date_format}{character. mdl mutation collect time column date format.}

\item{mdl_mutation_gene_col}{character. mdl sample information column for genename.}

\item{mdl_mutation_indication_col}{character. mdl sample information column for indication(like "wildtype","mutation").}

\item{mdl_mutations_aa_col}{character. mdl sample information column for amino acid}

\item{mdl_mutation_panel_col}{character. mdl sample information column for panel id.}

\item{mdl_mutation_keep_columns}{character. mdl mutation column kept for the output.}

\item{panel_info}{data.frame. panel information.}

\item{panel_id_col}{character. panel information column for panel id.}

\item{panel_chromosome_col}{character. panel information column for chromosome.}

\item{panel_start_col}{character. panel information column for start.}

\item{panel_end_col}{character. panel information column for end.}

\item{panel_gene_col}{character. panel information column for genename.}

\item{with_oncokb}{logic. annotated with oncoKB or not.}

\item{oncokb_db}{data.frame. oncoKB database.}

\item{oncokb_gene_col}{character. oncoKB column for genename}

\item{with_cosmic}{logic. annotated with cosmic or not.}

\item{cosmic_db}{data.frame. cosmic database.}

\item{cosmic_gene_col}{character. cosmic column for genename}

\item{cosmic_aa_col}{character. cosmic column for amino acid}

\item{mdl_columns2wes_columns}{character. a named character used for mapping column names of mdl mutations and wes(wgs) mutations}

\item{combined}{logic. if TRUE, consistent mutations between mdl mutations and wes/wgs mutations will be combined into one mutation.}
}
\value{
data.frame. annotated and merged mutations(WES/WGS,MDL)
}
\description{
annotate genomic mutations(WES/WGS) and mdl mutations
merge them if they are consistent
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
