% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_clustering.R
\name{map_clusters}
\alias{map_clusters}
\title{map_cluster}
\usage{
map_clusters(query_clusters, target_clusters, bestnumberofclusters)
}
\arguments{
\item{query_clusters}{integer vector.}

\item{target_clusters}{integer vector.}

\item{bestnumberofclusters}{integer vector, containing mapped cluster_id from target clusters}
}
\value{
integer vector, containing mapped cluster_id from target clusters
}
\description{
map cluster_id from query clusters to cluster_id from target clusters
}
\details{
Source provenance: GBM/XL_figures/vDotJun0823/GBMnat/R/funcsGBMnat.R (2025-08-11).
}
