% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_clustering.R
\name{multiCluster}
\alias{multiCluster}
\title{multiCluster}
\usage{
multiCluster(
  data,
  scale = c("row", "column", "none"),
  bestnumberofclusters,
  data_preTreat = F,
  removeVar = 0.2,
  variance_explained_cutoff = 0.8,
  nbclust_method = "complete",
  min.nc = 2,
  max.nc = 15,
  cluster_methods = c("kmeans", "pam", "hclust", "fuzzy", "mclust", "apclust", "hdbscan",
    "MCL", "specc", "kkmeans", "skmeans", "nmf", "som"),
  base_cluster_method
)
}
\arguments{
\item{data}{\code{data.frame()} or \code{matrix()}. Gene expression data (or matrix).}

\item{scale}{\code{character()}. Direction for scale data. Default ="row".}

\item{bestnumberofclusters}{\code{integer(1)}. Optimal number of clusters, if missing, calculated by estimate_bestNumberofClusters.}

\item{data_preTreat}{\code{logical()}. If TRUE, data will be PCA transformed; Components whose cumulative sum of variance reach variance_explained_cutoff will be kept. Default FALSE.}

\item{removeVar}{\code{numeric()}. Remove this percentage of variables based on low variance. Default=0.2.}

\item{variance_explained_cutoff}{\code{numeric()}. If data_preTreat is TRUE, components in PCA transformed data matrix whose cumulative sum of variance should reach this value. default=0.8.}

\item{nbclust_method}{the cluster analysis method to be used. This should be one of: "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid", "kmeans".}

\item{min.nc}{minimal number of clusters, between 1 and (number of objects - 1)}

\item{max.nc}{maximal number of clusters, between 2 and (number of objects - 1), greater or equal to min.nc. By default, max.nc=15.}

\item{cluster_methods}{\code{vector(mode="character")}. Methods used for clustering and evaluation. default=c("kmeans","pam","hclust","fuzzy","mclust","apcluster","hdbscan","MCL","specc","kkmeans","SKmeans","NMF","SOM").}

\item{base_cluster_method}{\code{character()}. if missing, the most correlated method will be used.}
}
\value{
list, containing best partition of sample, and co-clusters derived from multiple clustering algorithms.
}
\description{
estimate clustering using multiple clustering algorithms, such as "kmeans","pam","hclust","fuzzy","mclust","apclust","hdbscan","MCL","specc","kkmeans","SKmeans","NMF","SOM".
}
\details{
Source provenance: GBM/XL_figures/vDotJun0823/GBMnat/R/funcsGBMnat.R (2025-08-11).
}
\examples{
\dontrun{
multiCluster(data=data,scale='none',data_preTreat = T)
}

}
