% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mutation_analysis.R
\name{oncogenicpathway_tableplot}
\alias{oncogenicpathway_tableplot}
\title{oncogenicpathway_tableplot}
\usage{
oncogenicpathway_tableplot(
  pathways,
  mutations,
  sample_info,
  gene_column = "",
  sample_id_column = "Sample_ID",
  pathway_order,
  sample_order = NULL,
  top_anno,
  top_anno_sample_order,
  show_mutations = F,
  ...
)
}
\arguments{
\item{pathways}{list. no default, genesets for gsea, refer fgsea::gmtPathways}

\item{mutations}{data.frame. mutations data}

\item{sample_info}{data.frame. sample information.}

\item{gene_column}{character. mutation column for genename}

\item{sample_id_column}{character. sample information scolumn for sample id}

\item{pathway_order}{character. ordered pathway name.}

\item{sample_order}{character. ordered sample names.}

\item{top_anno}{HeatmapAnnotation. top annoation for oncogenic pathway table plot}

\item{top_anno_sample_order}{character. sample names for top annotation, consistent order with top annotaion}

\item{show_mutations}{logic. show mutation or not.}

\item{...}{list. extra parameter passed to ComplexHeatmap::Heatmap}
}
\value{
Heatmap. oncogenic pathway table plot
}
\description{
oncogenic pathway table plot
}
\details{
Source provenance: woodman_lab.XLi23/HaifengPackages.Mar2026/utiltools/R/mutations.R (2026-03-02).
}
