% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_spatial_immune_profiling.R
\name{performTtestsAllClassesOneVsRest}
\alias{performTtestsAllClassesOneVsRest}
\title{get t-test result}
\usage{
performTtestsAllClassesOneVsRest(dataMatrix, classVector, paired = FALSE, ...)
}
\arguments{
\item{dataMatrix}{The data matrix}

\item{classVector}{vector matching column names to categories}
}
\value{
A data frame of mean differences and p-values
}
\description{
get t-test result by group, one versus the rest
}
\details{
Source provenance: woodman_lab.XLi23/mIFvectra/R/vectraStatsFun.R (2024-07-02).
}
\examples{
t.result = performTtestsAllClassesOneVsRest(dataMatrix,classVector)
}
