% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_spatial_immune_profiling.R
\name{performTtestsAllRows}
\alias{performTtestsAllRows}
\title{get t-test result}
\usage{
performTtestsAllRows(dataGroup1, dataGroup2, ...)
}
\arguments{
\item{dataGroup1}{The group 1 data matrix. Feature by row, sample by column.}

\item{dataGroup2}{The group 2 data matrix. Feature by row, sample by column.}
}
\value{
A data frame of mean differences and p-values
}
\description{
perform profile-wise t-test by row
}
\details{
Source provenance: woodman_lab.XLi23/mIFvectra/R/vectraStatsFun.R (2024-07-02).
}
\examples{
t.result = performTtestsAllClassesOneVsRest(dataMatrix=data_mx,classVector=c("treatment1","treatment2","treatment1","control"));
}
