% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_survival_analysis.R
\name{plotKM}
\alias{plotKM}
\title{generate Kaplan-Meier survival curve}
\usage{
plotKM(
  survdf,
  duration,
  censoring,
  variable = NULL,
  conversion = "d2m",
  statOnly = F,
  palette = NULL,
  OrderedVarFactor = NULL,
  plotTitle = NULL,
  table_base_size = 6,
  timeUnit = "month",
  timeYLab = "OS"
)
}
\arguments{
\item{survdf}{\code{data.frame()}.
Should at least include columns of duration (eg. survival), censoring (eg. vital status), grouping varible.}

\item{duration}{\code{character()}. The name of the column of survival.}

\item{censoring}{\code{character()}. The name of the column of vital status.}

\item{variable}{\code{character()}. The name of the column of grouping variable.}

\item{conversion}{if "d2m". Convert duration unit from days to months.}

\item{palette}{\code{character()}. color palette. character vector.}

\item{OrderedVarFactor}{\code{character()}. order of the factors in grouping variable.}

\item{plotTitle}{\code{character()}. plot title.}

\item{table_base_size}{annotation font size.}
}
\value{
a Kaplan-Meier plot.
}
\description{
generate Kaplan-Meier survival curve
}
\details{
Source provenance: GBM/XL_figures/vDotJun0823/GBMnat/R/funcsGBMnat.R (2025-08-11).
}
