% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_spatial_immune_profiling.R
\name{robustscale}
\alias{robustscale}
\title{robust scaling
uses median an mad instead of mean and row
applies the scaling to the columns (samples) by default}
\usage{
robustscale(
  data,
  dim = 1,
  center = TRUE,
  scale = TRUE,
  scale_max = Inf,
  preserveScale = TRUE
)
}
\arguments{
\item{data}{matrix or data.frame}

\item{dim}{should rows (1) or columns (2:default) be scaled}

\item{center}{subract median (default:TRUE)}

\item{scale}{scale by mad  (default:TRUE)}

\item{preserveScale}{default TRUE , equalize scales but do not change them}
}
\description{
robust scaling
uses median an mad instead of mean and row
applies the scaling to the columns (samples) by default
}
\details{
Source provenance: woodman_lab.XLi23/mIFvectra/R/vectraPlotsFun.R (2024-04-09).
}
\examples{
library(quantable)
tmp = matrix(rep((1:100),times = 4) + rnorm(100*4,0,3),ncol=4)
mean = c(20,30,10,40)
sd = c(4,3,4,5)
tmp = sweep(tmp,2,sd,"*")
tmp = sweep(tmp,2,mean,"+")
boxplot(tmp)
tmp = robustscale(tmp)
boxplot(tmp$data)
}
