% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_expression_analysis.R
\name{unsupervised_analysis}
\alias{unsupervised_analysis}
\title{unsupervised_analysis}
\usage{
unsupervised_analysis(
  dat,
  labels = NULL,
  scaled = c("row", "column", "none"),
  run_umap = T,
  umap_params = NULL,
  umap_config = umap::umap.defaults,
  run_tsne = T,
  tsne_params = NULL,
  run_pca = T,
  pca_params = NULL,
  run_mds = T,
  mds_params = NULL,
  run_heatmap = T,
  heatmap_params = NULL,
  cols
)
}
\arguments{
\item{dat}{data.frame. data for unsupervised analysis.}

\item{labels}{character. labels for samples.}

\item{scaled}{character. scale direction.}

\item{run_umap}{logic. run umap method or not.}

\item{umap_params}{list. contain parameters for umap.}

\item{umap_config}{list. contain configuration for umap.}

\item{run_tsne}{logic. run umap method or not.}

\item{tsne_params}{list. contain parameters for tsne.}

\item{run_pca}{logic. run umap method or not.}

\item{pca_params}{list. contain parameters for pca}

\item{run_mds}{logic. run umap method or not.}

\item{mds_params}{list. contain parameters for mds.}

\item{run_heatmap}{logic. run umap method or not.}

\item{heatmap_params}{list. contain parameters for heatmap.}

\item{cols}{character. colors for labels.}
}
\value{
list. contain results for each method.
}
\description{
unsupervised analysis of expression
}
\details{
Source provenance: mRNA.R (2026-03-02).
}
